requirements.txt
streamlit
google-generativeai
python-dotenv
PyPdf2  
db-sqlite3